[review] Fix Sonar warnings

+++ test/device/PAC_dev_tests.cpp
@@ -1847,14 +1847,14 @@ TEST( level_e_cyl_hor, get_volume_half_full )
     level_e_cyl_hor test_dev( "test_LT1" );
 
     // Set R=1m, L=2m, density=1000 kg/m³, no angle, no offset.
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_R,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_R ),
         test_dev.start_param_idx, 1.0f );
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_H,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_H ),
         test_dev.start_param_idx, 2.0f );
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_DENSITY,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_DENSITY ),
         test_dev.start_param_idx, 1000.0f );
     // P_MAX_P = 0.1962 bar → h = 0.5 * P_MAX_P * 1e5 / (rho*g) = 1m = R.
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_MAX_P,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_MAX_P ),
         test_dev.start_param_idx, 0.1962f );
     // 50% fill → pressure = 0.0981 bar → h = 1m = R (half full).
     test_dev.direct_set_value( 50.0f );
@@ -1868,14 +1868,14 @@ TEST( level_e_cyl_hor, get_volume_full )
     {
     level_e_cyl_hor test_dev( "test_LT1" );
 
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_R,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_R ),
         test_dev.start_param_idx, 1.0f );
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_H,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_H ),
         test_dev.start_param_idx, 2.0f );
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_DENSITY,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_DENSITY ),
         test_dev.start_param_idx, 1000.0f );
     // P_MAX_P = 0.4 bar → h = P_MAX_P * 1e5 / (rho*g) ≈ 4 m > 2R at 100%.
-    test_dev.set_par( level_e_cyl_hor::CONSTANTS::P_MAX_P,
+    test_dev.set_par( static_cast<int>( level_e_cyl_hor::CONSTANTS::P_MAX_P ),
         test_dev.start_param_idx, 0.4f );
     test_dev.direct_set_value( 100.0f );
